Beginning of darwin SimOS; Still quite rough around the edges
Need to patch up all of the loader nonsense, including the weird shared
memory region that appears to just be at a forever static address in
macOS. Need to add the Mach syscall table. Need to properly grab symbols
from dyld, and also set up `dyld` as part of the loader, just as
there's already `ld`.